requirements.txt
absl-py==1.3.0
anyio==3.6.2
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.2.0
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
bokeh==3.0.3
Brlapi==0.8.3
cachetools==5.2.0
canonicaljson==1.6.4
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
cnab==0.1.7
colorama==0.4.4
comm==0.1.1
command-not-found==0.3
contourpy==1.0.6
cryptography==3.4.8
cubic==0.5.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.4
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
dj-static==0.0.6
Django==4.1.5
django-stdimage==6.0.1
dnspython==2.1.0
docutils==0.19
dotenv-cli==3.1.0
duplicity==0.8.21
entrypoints==0.4
et-xmlfile==1.1.0
Everything-Tkinter==0.2
exceptiongroup==1.1.0
executing==1.2.0
fastapi==0.88.0
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.8.2
Flask==2.2.2
flatbuffers==22.11.23
fonttools==4.38.0
future==0.18.2
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.1
gunicorn==20.1.0
gyp==0.1
h11==0.13.0
h5py==3.7.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.18.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonify==0.5
jsonschema==4.17.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keeper-cnab240==0.0b11
keras==2.11.0
keyring==23.5.0
Kivy==2.1.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.6
load-dotenv==0.1.0
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mediapipe==0.9.0
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
multitasking==0.0.11
nbclassic==0.4.8
nbclient==0.7.1
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.11.0
neuralintents==0.0.4
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
peewee==3.15.4
pefile==2021.9.3
pewee==0.1.0.dev9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==3.12.4
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.2
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
pypokedex==1.6.0
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.19.2
PyScreeze==0.1.28
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
python-networkdays==1.1
python3-cnab==1.0.11
python3-xlib==0.15
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
random2==1.0.1
regex==2022.10.31
reportlab==3.6.8
requests==2.28.1
requests-oauthlib==1.3.1
requires.io==0.2.6
rsa==4.9
scipy==1.9.3
seaborn==0.12.1
SecretStorage==3.3.1
selenium==4.7.2
Send2Trash==1.8.0
setuptools-git==1.1
simplejson==3.18.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.44
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.22.0
static3==0.7.0
systemd-python==234
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
terminado==0.17.0
tinycss2==1.2.1
tk==0.1.0
tkinter-math==0.2.4
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
trio==0.22.0
trio-websocket==0.9.2
ttystatus==0.38
typing_extensions==4.4.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.15.0
vboxapi==1.0
virtualenv==20.17.1
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
wsproto==1.0.0
xdg==5
xkit==0.0.0
xyzservices==2023.2.0
yfinance==0.1.87
zipp==1.0.0
